refactor(mcp): extract notes tools to tools_notes.py
